+++ FishSpinDays/FishSpinDays.Web/Helpers/Filters/ApiMetricsAttribute.cs
@@ -1,6 +1,5 @@
 ﻿namespace FishSpinDays.Web.Helpers.Filters
 {
-    using FishSpinDays.Web.Controllers;
     using Microsoft.AspNetCore.Mvc.Filters;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Logging;
@@ -14,7 +13,7 @@
     {
         public override async Task OnActionExecutionAsync(ActionExecutingContext context, ActionExecutionDelegate next)
         {
-            var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<PublicationsAPIController>>();
+            var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<ApiMetricsAttribute>>();
 
             var endpoint = context.ActionDescriptor.DisplayName;
             var method = context.HttpContext.Request.Method;
