+++ FishSpinDays/FishSpinDays.Web/Helpers/Filters/ImportantOperationAttribute.cs
@@ -1,6 +1,5 @@
 ﻿namespace FishSpinDays.Web.Helpers.Filters
 {
-    using FishSpinDays.Web.Controllers;
     using Microsoft.AspNetCore.Mvc.Filters;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Logging;
@@ -16,7 +15,7 @@
 
         public override async Task OnActionExecutionAsync(ActionExecutingContext context, ActionExecutionDelegate next)
         {
-            var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<PublicationsAPIController>>();
+            var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<ImportantOperationAttribute>>();
             var operationName = OperationName ?? context.ActionDescriptor.DisplayName;
 
             using var scope = logger.BeginScope("Important Operation: {Operation}", operationName);
