+++ FishSpinDays/FishSpinDays.Web/Helpers/Filters/SimpleOperationAttribute.cs
@@ -1,6 +1,5 @@
 ﻿namespace FishSpinDays.Web.Helpers.Filters
 {
-    using FishSpinDays.Web.Controllers;
     using Microsoft.AspNetCore.Mvc.Filters;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Logging;
@@ -15,7 +14,7 @@
 
         public override async Task OnActionExecutionAsync(ActionExecutingContext context, ActionExecutionDelegate next)
         {
-            var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<PublicationsAPIController>>();
+            var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<SimpleOperationAttribute>>();
             var operationName = OperationName ?? context.ActionDescriptor.DisplayName;
 
             logger.LogDebug("Simple API operation: {Operation}", operationName);
